[open][neva_upstreamable][app_runtime][error_page] Disable error page site isolation
:Release Notes:
Disable error page site isolation.

:Detailed Notes:
Neva app runtime is still not ready for site isolation or error page
site isolation. We found that at least the WebkitPreferences are not
propagated to the RenderView created for error page.

Upstream policy has changed from disabling by default on content API
the site isolation for error page, to enable it if the error page
replaces the main frame. And the WebkitPreferences in app_runtime::WebView
were not updated accordingly.

As we do not know if there are other side effects we do not know
for enabling error page site isolation, we disable it for now as
we did for general site isolation.

:Testing Performed:
1. restart webapp-mgr
2. Launch viewster, check the PID (the only WebAppMgr with type=renderer).
3. Connect through web inspector and side load a non existing web page
  so it shows the error page
4. Check the only WebAppMgr process with type=renderer keeps being
  the same PID.

:QA Notes:
Test-cases: GT-27589

:Issues Addressed:
[PLAT-80044] Port "[closed][neva_upstreamable][a11y] Read network error text when app doesn't support audioguidance"
[PLAT-80534] [chromium72] Error page should reuse web app renderer process

Change-Id: Ie715857d84d6174f4985067d81651adf03d41ab7
Reviewed-on: http://gpro.lge.com/244811