fixup! [open][neva_upstreamable][keep_alive] Implement keep alive feature
:Release Notes:
Send Widget Close IPC message with right routing

:Detailed Notes:
As part of the work for keep alive, the method RenderWidget::DoDeferredClose
was made virtual, and overriden in RenderViewImpl. It was still
sending the method WidgetHostMsg_Close through IPC.

The problem was that the routing id of RenderViewImpl and RenderWidget were not
the same, so the listener in browser process of the call was wrong.

To avoid this, as DoDeferredClose is private (so we cannot just call the parent
class implementation), this patch changes the approach. It just removes virtual
of DoDeferredClose again, but calls a new public virtual named
WillDoDeferredClose, that can be overriden in RenderViewImpl. This way the
call to WidgetHostMsg_Close is done in RenderWidget with the right routing id.

:Testing Performed:
1. Long-press settings button -> settings app is visible.
2. Clock on the right half of the screen to close settings -> settings now closes.

:QA Notes:
Test-cases: GT-27610

:Issues Addressed:
[PLAT-80742] [chromium72] Settings app doesn't close when clicked outside

Change-Id: I47ff4a89cc87abbb1b42bfc318e8bf0e5f166871
Reviewed-on: http://gpro.lge.com/245420