Revert "[open][upstreamable][threading] Do not stash thread id on TLS in Linux ARM32 if built with GCC"
:Release Notes:
Revert of fix for PlatformThread::CurrentId in ARM32+GCC

:Detailed Notes:
This reverts commit 8c688aca750f0e33dd011da87fd9931c38023881.

After some discussion upstream and investigation downstream, we found the
root cause of the TLS corruption that was breaking PlatformThread::CurrentId.

This was caused by tcmalloc ThreadCache storing a pointer to its heap in TLS
using initial-exec TLS model, that is broken in ARM32. A new patch has been
proposed upstream that will be sent as a follow up.

:Testing Performed:

:QA Notes:
Test-cases: GT-24343

:Issues Addressed:
[PLAT-81106] [chromium72] 'CrRenderMain / CrBrowserMain' displayed instead of 'WebAppMgr' on 'top' command.
[PLAT-80354] [chromium72] PlatformThread::CurrentId() does not return the thread ID, but some weird value

Change-Id: I563f24241ee435d1c1978b94aad586ba4ade8858
Originally-Reviewed-on: http://gpro.lge.com/248254
Reviewed-on: http://gpro.lge.com/248459
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>